Relocates RTModule and RTScope into separate headers and compilation units.
